2.43.3016     ★★公布、添加功能         添加 Rafy.Web.Http、Rafy.Web.Mvc 两个命名空间。（代码移植自历史项目）。     添加接口         AggtSerializer 在序列化枚举时，可以选择使用 字符串 或 Label 进行输出。         OData 支持枚举属性的查询。         添加 CloneOptions.NewSingleEntity();     添加功能         在输出 Sql 日志时，如果是 SQLServer 数据库，还同时输出对应的连接的客户端 Id。     BUG修复         Oracle 支持问题。         解决 WebOrThreadStaticAppContextProvider 中的代码引发 ThreadStaticAppContextProvider.Items 集合出错的问题。         解决‘无法更新枚举类型的冗余属性。’的 BUG。
2.42.2840
    ★★公布、添加功能
        添加幽灵实体插件，实现实体的假删除功能。
    添加接口
        为 RepositoryDataProvider 添加删除前、查询前两个事件扩展点。
        添加 WebAppStarter。
        添加 ExtensionPropertiesRegister。
    重构
        DataSaver.Delete 改为 DeleteFromPersistence。

2.41.2815
    ★★公布、添加功能
        支持使用实体类中的注释来更新数据库中的相关注释内容。
            实现在 SqlServer 中生成注释。
            实现在 Oracle 中生成注释。
        Json.AggtSerializer 支持序列化、反序列化树型实体。
        DbAccesserFactory 支持从仓库创建 IDbAccesser。
    BUG修复
        解决使用 Like 语句模糊查询时，参数值中有通配符 % 及 _ 时无效的问题。
        解决 ORACLE 中把列从可空变为不可空时，更新空列的默认时间时使用字符串，导致语法错误的问题。

+++ Rafy/Plugins/Rafy.Domain.ORM.BatchSubmit/BatchImporter.cs
@@ -17,6 +17,7 @@ using System.Data;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Rafy.Data;
 using Rafy.Domain.Caching;
 
 namespace Rafy.Domain.ORM.BatchSubmit
@@ -140,12 +141,20 @@ namespace Rafy.Domain.ORM.BatchSubmit
         /// <param name="batch"></param>
         protected virtual void ImportDelete(EntityBatch batch)
         {
-            string sql = "DELETE FROM " + batch.Table.Name + " WHERE ID IN (";
-            bool needDelimiter = false;
-
             foreach (var section in this.EnumerateAllBatches(batch.DeleteBatch, 1000))
             {
-                var sqlDelete = new StringBuilder(sql);
+                FormattedSql sqlDelete = null;
+                if (batch.Repository.EntityMeta.IsPhantomEnabled)
+                {
+                    sqlDelete = "UPDATE " + batch.Table.Name + " SET ISPHANTOM = {0} WHERE ID IN (";
+                    sqlDelete.Parameters.Add(BooleanBoxes.True);
+                }
+                else
+                {
+                    sqlDelete = "DELETE FROM " + batch.Table.Name + " WHERE ID IN (";
+                }
+
+                bool needDelimiter = false;
                 for (int i = 0, c = section.Count; i < c; i++)
                 {
                     var item = section[i];
@@ -167,7 +176,7 @@ namespace Rafy.Domain.ORM.BatchSubmit
                 }
                 sqlDelete.Append(')');
 
-                batch.DBA.ExecuteText(sqlDelete.ToString());
+                batch.DBA.ExecuteText(sqlDelete, sqlDelete.Parameters);
             }
         }
